improved AsLeader()
- updated to go 1.18
- now possible to run multiple `AsLeader` per natty instance

... but I think it should actually be updated more and allow AsLeader to run
across multiple instances against the same bucket/key.

This means that state shouldn't be saved or checked against natty instance itself
but via NATS instead. That'll *definitely* be better and probably what is expected.

Will do that tomorrow.

This work is needed for `monitor` rate check as it makes significant use of AsLeader